+++ BancoMedellin/Server/Services/UsuarioService/IUsuarioService.cs
@@ -4,9 +4,8 @@
     public interface IUsuarioService
     {
         Task<List<Usuario>> GetAll();
-        Task<Usuario> GetUsuarioByDni(ulong Dni);
+        Task<Usuario> GetUsuarioByDni(int Dni);
         Task<Usuario> Registrar(UsuarioDTO usuario);
         Task<string> Login(UsuarioDTO usuario);
-        
     }
 }
