+++ BancoMedellin/Shared/Usuario.cs
@@ -11,7 +11,7 @@ namespace BancoMedellin.Shared
     public class Usuario
     {
         [Key, DatabaseGenerated(DatabaseGeneratedOption.None)]
-        public ulong Dni { get; set; }
+        public int Dni { get; set; }
         [Required, MaxLength(64)]
         public string Nombre { get; set; } = string.Empty;
